Updates to entities made during class

+++ src/main/java/org/capstone/repository/LeaveRepository.java
@@ -1,8 +1,7 @@
 package org.capstone.repository;
 
 import org.capstone.entity.Leave;
-import org.capstone.entity.SiteUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeaveRepository extends JpaRepository<Leave, Long> {
+public interface LeaveRepository extends JpaRepository<Leave, Integer> {
 }
